# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: TheLuxApp AI ChatBot API
  version: 1.0.0
  description: |
    Enterprise AI chatbot platform API. Multi-tenant, multi-model, with RAG,
    tool execution, memory, and platform integrations (Discord, Slack, Telegram,
    Teams, WhatsApp, web embed).

    **Authentication:**
    - `Bearer <JWT>` — from `POST /auth/token`
    - `ApiKey <key>` — from `POST /auth/api-keys` (requires `X-Tenant-ID` header)

  contact:
    name: TheLuxApp GitHub
    url: https://github.com/iboss21/TheLuxApp-Ai-ChatBot
  license:
    name: MIT
servers:
  - url: https://your-domain.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: `ApiKey lux_xxxxxxxx`"
  schemas:
...
    Integration:
      type: object
      properties:
        id:         { type: string, format: uuid }
        platform:   { type: string, enum: [discord, slack, telegram, teams, whatsapp] }
        name:       { type: string }
        enabled:    { type: boolean }
        created_at: { type: string, format: date-time }
security:
  - BearerAuth: []

# ════════════════════════════════════════════════════════════════════════════
paths:
...
  /v1/integrations/manage:
    get:
      tags: [Integrations]
      summary: List platform integrations (admin)
      responses:
        "200":
          description: Integration list
          content:
            application/json:
              schema:
                type: object
                properties:
                  integrations:
                    type: array
                    items: { $ref: "#/components/schemas/Integration" }

    post:
      tags: [Integrations]
      summary: Register a platform integration (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, name]
              properties:
                platform:    { type: string, enum: [discord, slack, telegram, teams, whatsapp] }
                name:        { type: string }
                config:      { type: object, description: "Platform-specific config (bot_token, signing_secret, etc.)" }
                bot_user_id: { type: string, format: uuid }
            examples:
              discord:
                value:
                  platform: discord
                  name: "My Discord Server"
                  config:
                    application_id: "123456"
                    bot_token: "Bot xxx"
                    public_key: "abc123"
              slack:
                value:
                  platform: slack
                  name: "My Slack Workspace"
                  config:
                    bot_token: "xoxb-..."
                    signing_secret: "abc..."
              telegram:
                value:
                  platform: telegram
                  name: "My Telegram Bot"
                  config:
                    bot_token: "123456:ABCdef"
      responses:
        "201":
          description: Integration registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Integration" }
...
  /v1/integrations/discord/webhook:
    post:
      tags: [Integrations]
      summary: Discord interactions endpoint
      security: []
      description: |
        Receives Discord slash command interactions. Verified via Ed25519 signature.

        **Required headers:**
        - `X-Signature-Ed25519`: signature from Discord
        - `X-Signature-Timestamp`: timestamp from Discord
      parameters:
        - { name: X-Signature-Ed25519,  in: header, required: true, schema: { type: string } }
        - { name: X-Signature-Timestamp, in: header, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:  { type: integer, description: "1=PING, 2=APPLICATION_COMMAND" }
                token: { type: string }
                data:
                  type: object
                  properties:
                    name: { type: string }
                    options:
                      type: array
                      items:
                        type: object
                        properties:
                          name:  { type: string }
                          value: { type: string }
      responses:
        "200":
          description: Interaction response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type: { type: integer, description: "1=PONG, 4=CHANNEL_MESSAGE, 5=DEFERRED" }
        "401": { description: Invalid signature }
  /v1/integrations/slack/events:
    post:
      tags: [Integrations]
      summary: Slack Events API endpoint
      security: []
      description: |
        Receives Slack events (messages, mentions). Verified via HMAC-SHA256.

        **Required headers:**
        - `X-Slack-Signature`: HMAC signature
        - `X-Slack-Request-Timestamp`: Unix timestamp
      parameters:
        - { name: X-Slack-Signature,         in: header, required: true, schema: { type: string } }
        - { name: X-Slack-Request-Timestamp,  in: header, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:      { type: string, enum: [url_verification, event_callback] }
                challenge: { type: string, description: "Present for url_verification" }
                event:
                  type: object
                  properties:
                    type:    { type: string }
                    user:    { type: string }
                    text:    { type: string }
                    channel: { type: string }
      responses:
        "200":
          description: OK (or challenge for url_verification)
        "401": { description: Invalid signature }
  /v1/integrations/telegram/{token}/webhook:
    post:
      tags: [Integrations]
      summary: Telegram bot webhook
      security: []
      description: Receives Telegram updates. The bot token in the URL path serves as authentication.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string }, description: "Telegram bot token" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                update_id: { type: integer }
                message:
                  type: object
                  properties:
                    from: { type: object, properties: { id: { type: integer }, username: { type: string } } }
                    chat: { type: object, properties: { id: { type: integer } } }
                    text: { type: string }
      responses:
        "200": { description: OK }
        "401": { description: Invalid token }
  /v1/integrations/teams/webhook:
    post:
      tags: [Integrations]
      summary: Microsoft Teams Bot Framework webhook
      security: []
      description: Receives Bot Framework Activity objects from Microsoft Teams.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:        { type: string, enum: [message] }
                text:        { type: string }
                from:        { type: object }
                conversation: { type: object }
                serviceUrl:  { type: string }
      responses:
        "200":
          description: Bot Framework reply Activity
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:      { type: string }
                  text:      { type: string }
                  replyToId: { type: string }
  /v1/integrations/whatsapp/webhook:
    get:
      tags: [Integrations]
      summary: WhatsApp webhook verification challenge
      security: []
      parameters:
        - { name: "hub.mode",         in: query, schema: { type: string } }
        - { name: "hub.verify_token", in: query, schema: { type: string } }
        - { name: "hub.challenge",    in: query, schema: { type: string } }
      responses:
        "200": { description: Challenge string }
        "403": { description: Verification failed }

    post:
      tags: [Integrations]
      summary: WhatsApp Cloud API webhook
      security: []
      description: Receives WhatsApp message events from Meta Cloud API.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                object: { type: string, enum: [whatsapp_business_account] }
                entry:  { type: array }
      responses:
        "200": { description: EVENT_RECEIVED }

  # ── ADMIN ─────────────────────────────────────────────────────────────────
...